⏪️ revert : 통합테스트 관리 논의  필요
