+++ src/test/java/com/devcourse/checkmoi/template/IntegrationTest.java
@@ -6,19 +6,12 @@ import static org.springframework.restdocs.headers.HeaderDocumentation.requestHe
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import com.devcourse.checkmoi.domain.book.repository.BookRepository;
-import com.devcourse.checkmoi.domain.comment.repository.CommentRepository;
-import com.devcourse.checkmoi.domain.post.repository.PostRepository;
-import com.devcourse.checkmoi.domain.study.repository.StudyMemberRepository;
-import com.devcourse.checkmoi.domain.study.repository.StudyRepository;
 import com.devcourse.checkmoi.domain.token.dto.TokenResponse.TokenWithUserInfo;
-import com.devcourse.checkmoi.domain.user.repository.UserRepository;
 import com.devcourse.checkmoi.global.security.oauth.OAuthService;
 import com.devcourse.checkmoi.global.security.oauth.UserProfile;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.UUID;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -43,23 +36,6 @@ public abstract class IntegrationTest {
     @Autowired
     protected OAuthService oAuthService;
 
-    @Autowired
-    protected CommentRepository commentRepository;
-
-    @Autowired
-    protected UserRepository userRepository;
-
-    @Autowired
-    protected StudyRepository studyRepository;
-
-    @Autowired
-    protected BookRepository bookRepository;
-
-    @Autowired
-    protected StudyMemberRepository studyMemberRepository;
-
-    @Autowired
-    protected PostRepository postRepository;
 
     private static UserProfile createUserProfile(String name) {
         return UserProfile.builder()
@@ -95,13 +71,4 @@ public abstract class IntegrationTest {
         return oAuthService.register(createUserProfile(UUID.randomUUID().toString().substring(19)));
     }
 
-    @AfterEach
-    void cleanUp() {
-        commentRepository.deleteAllInBatch();
-        postRepository.deleteAllInBatch();
-        studyMemberRepository.deleteAllInBatch();
-        studyRepository.deleteAllInBatch();
-        bookRepository.deleteAllInBatch();
-        userRepository.deleteAllInBatch();
-    }
 }
